backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --format requirements-txt --output-file testrequirements.txt --no-hashes
annotated-types==0.7.0
    # via pydantic
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
django==5.2.4
    # via
    #   backend
    #   django-cors-headers
    #   django-ninja
django-cors-headers==4.7.0
    # via backend
django-ninja==1.4.3
    # via backend
psycopg2-binary==2.9.10
    # via backend
pydantic==2.11.7
    # via django-ninja
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via backend
sqlparse==0.5.3
    # via django
typing-extensions==4.14.1
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via
    #   django
    #   pandas
whitenoise==6.9.0
    # via backend
pandas==2.3.1
    # via scraper
gunicorn==23.0.0
    # via backend
numpy==2.3.1
    # via pandas
packaging==25.0
    # via gunicorn
python-dateutil==2.9.0.post0
    # via pandas
six==1.17.0
    # via python-dateutil
pytz==2025.2
    # via pandas
torch==2.7.1
textblob==0.19.0
nltk==3.9.1
scipy==1.16.0
scikit-learn==1.7.0
